+++ src/MyPluginInfo.cs
@@ -2,9 +2,9 @@
 {
     public static class MyPluginInfo
     {
-        public const string PLUGIN_GUID = "Scopophobia";
+        public const string PLUGIN_GUID = "stysk1.Scopophobia";
 
         public const string PLUGIN_NAME = "Scopophobia";
-        public const string PLUGIN_VERSION = "1.3.2";
+        public const string PLUGIN_VERSION = "2.0.0";
     }
 }